[IFC][SVG text] Move layout code out from legacy inline boxes https://bugs.webkit.org/show_bug.cgi?id=283792 rdar://problem/140651613
Reviewed by Alan Baradlay.

Reduce dependencies to legacy inline boxes.

* Source/WebCore/layout/integration/inline/InlineIteratorSVGTextBox.cpp:
* Source/WebCore/rendering/LegacyLineLayout.cpp:
(WebCore::LegacyLineLayout::createLineBoxesFromBidiRuns):

RenderSVGText::layout() now triggers computePerCharacterLayoutInformation() directly.

* Source/WebCore/rendering/svg/RenderSVGText.cpp:
(WebCore::RenderSVGText::layout):
(WebCore::RenderSVGText::computePerCharacterLayoutInformation):
(WebCore::RenderSVGText::layoutCharactersInTextBoxes):
(WebCore::RenderSVGText::layoutChildBoxes):
(WebCore::RenderSVGText::layoutRootBox):
(WebCore::swapItemsInLayoutAttributes):
(WebCore::findFirstAndLastAttributesInVector):
(WebCore::reverseInlineBoxRangeAndValueListsIfNeeded):
(WebCore::RenderSVGText::reorderValueListsToLogicalOrder):
(WebCore::RenderSVGText::positionForPoint):
(WebCore::RenderSVGText::legacyRootBox const):
* Source/WebCore/rendering/svg/RenderSVGText.h:
* Source/WebCore/rendering/svg/SVGRootInlineBox.cpp:
(WebCore::SVGRootInlineBox::computePerCharacterLayoutInformation): Deleted.
(WebCore::SVGRootInlineBox::layoutCharactersInTextBoxes): Deleted.
(WebCore::SVGRootInlineBox::layoutChildBoxes): Deleted.
(WebCore::SVGRootInlineBox::layoutRootBox): Deleted.
(WebCore::swapItemsInLayoutAttributes): Deleted.
(WebCore::findFirstAndLastAttributesInVector): Deleted.
(WebCore::reverseInlineBoxRangeAndValueListsIfNeeded): Deleted.
(WebCore::SVGRootInlineBox::reorderValueListsToLogicalOrder): Deleted.
* Source/WebCore/rendering/svg/SVGRootInlineBox.h:

Canonical link: https://commits.webkit.org/287151@main
