+++ Source/WebCore/rendering/svg/RenderSVGText.cpp
@@ -33,6 +33,7 @@
 #include "HitTestRequest.h"
 #include "HitTestResult.h"
 #include "InlineIteratorBoxInlines.h"
+#include "InlineIteratorLogicalOrderTraversal.h"
 #include "InlineIteratorSVGTextBox.h"
 #include "LayoutIntegrationLineLayout.h"
 #include "LayoutRepainter.h"
@@ -47,6 +48,9 @@
 #include "RenderSVGInlineText.h"
 #include "RenderSVGRoot.h"
 #include "SVGElementTypeHelpers.h"
+#include "SVGInlineFlowBox.h"
+#include "SVGInlineTextBox.h"
+#include "SVGInlineTextBoxInlines.h"
 #include "SVGLengthList.h"
 #include "SVGRenderStyle.h"
 #include "SVGRenderingContext.h"
@@ -54,6 +58,7 @@
 #include "SVGRootInlineBox.h"
 #include "SVGTextBoxPainter.h"
 #include "SVGTextElement.h"
+#include "SVGTextLayoutEngine.h"
 #include "SVGURIReference.h"
 #include "SVGVisitedRendererTracking.h"
 #include "TransformState.h"
@@ -402,6 +407,8 @@ void RenderSVGText::layout()
     rebuildFloatingObjectSetFromIntrudingFloats();
     layoutInlineChildren(true, repaintLogicalTop, repaintLogicalBottom);
 
+    computePerCharacterLayoutInformation();
+
     // updatePositionAndOverflow() is called by SVGRootInlineBox, after forceLayoutInlineChildren() ran, before this point is reached.
     if (m_needsReordering)
         m_needsReordering = false;
@@ -433,6 +440,206 @@ void RenderSVGText::layout()
     repainter.repaintAfterLayout();
     clearNeedsLayout();
     m_hasPerformedLayout = true;
+}
+
+void RenderSVGText::computePerCharacterLayoutInformation()
+{
+    if (!hasLines())
+        return;
+
+    if (m_layoutAttributes.isEmpty())
+        return;
+
+    if (m_needsReordering)
+        reorderValueListsToLogicalOrder();
+
+    // Perform SVG text layout phase two (see SVGTextLayoutEngine for details).
+    SVGTextLayoutEngine characterLayout(m_layoutAttributes);
+    layoutCharactersInTextBoxes(legacyRootBox(), characterLayout);
+
+    // Perform SVG text layout phase three (see SVGTextChunkBuilder for details).
+    auto fragmentMap = characterLayout.finishLayout();
+
+    // Perform SVG text layout phase four
+    // Position & resize all SVGInlineText/FlowBoxes in the inline box tree, resize the root box as well as the RenderSVGText parent block.
+    auto childRect = layoutChildBoxes(legacyRootBox(), fragmentMap);
+    layoutRootBox(childRect);
+}
+
+void RenderSVGText::layoutCharactersInTextBoxes(LegacyInlineFlowBox* start, SVGTextLayoutEngine& characterLayout)
+{
+    for (auto* child = start->firstChild(); child; child = child->nextOnLine()) {
+        if (auto* legacyTextBox = dynamicDowncast<SVGInlineTextBox>(*child)) {
+            ASSERT(is<RenderSVGInlineText>(legacyTextBox->renderer()));
+            characterLayout.layoutInlineTextBox(InlineIterator::svgTextBoxFor(legacyTextBox));
+        } else {
+            // Skip generated content.
+            RefPtr node = child->renderer().node();
+            if (!node)
+                continue;
+
+            auto& flowBox = downcast<SVGInlineFlowBox>(*child);
+            bool isTextPath = node->hasTagName(SVGNames::textPathTag);
+            if (isTextPath) {
+                // Build text chunks for all <textPath> children, using the line layout algorithm.
+                // This is needeed as text-anchor is just an additional startOffset for text paths.
+                SVGTextLayoutEngine lineLayout(characterLayout.layoutAttributes());
+                layoutCharactersInTextBoxes(&flowBox, lineLayout);
+
+                characterLayout.beginTextPathLayout(downcast<RenderSVGTextPath>(child->renderer()), lineLayout);
+            }
+
+            layoutCharactersInTextBoxes(&flowBox, characterLayout);
+
+            if (isTextPath)
+                characterLayout.endTextPathLayout();
+        }
+    }
+}
+
+FloatRect RenderSVGText::layoutChildBoxes(LegacyInlineFlowBox* start, SVGTextFragmentMap& fragmentMap)
+{
+    FloatRect childRect;
+
+    for (auto* child = start->firstChild(); child; child = child->nextOnLine()) {
+        FloatRect boxRect;
+        if (auto* textBox = dynamicDowncast<SVGInlineTextBox>(*child)) {
+            ASSERT(is<RenderSVGInlineText>(textBox->renderer()));
+
+            auto it = fragmentMap.find(makeKey(*InlineIterator::svgTextBoxFor(textBox)));
+            if (it != fragmentMap.end())
+                textBox->setTextFragments(WTFMove(it->value));
+
+            boxRect = textBox->calculateBoundaries();
+            textBox->setX(boxRect.x());
+            textBox->setY(boxRect.y());
+            textBox->setLogicalWidth(boxRect.width());
+            textBox->setLogicalHeight(boxRect.height());
+        } else {
+            // Skip generated content.
+            if (!child->renderer().node())
+                continue;
+
+            auto& flowBox = downcast<SVGInlineFlowBox>(*child);
+            layoutChildBoxes(&flowBox, fragmentMap);
+
+            boxRect = flowBox.calculateBoundaries();
+            flowBox.setX(boxRect.x());
+            flowBox.setY(boxRect.y());
+            flowBox.setLogicalWidth(boxRect.width());
+            flowBox.setLogicalHeight(boxRect.height());
+        }
+        childRect.unite(boxRect);
+    }
+
+    return childRect;
+}
+
+void RenderSVGText::layoutRootBox(const FloatRect& childRect)
+{
+    // Finally, assign the root block position, now that all content is laid out.
+    updatePositionAndOverflow(childRect);
+
+    // Position all children relative to the parent block.
+    for (auto* child = legacyRootBox()->firstChild(); child; child = child->nextOnLine()) {
+        // Skip generated content.
+        if (!child->renderer().node())
+            continue;
+        child->adjustPosition(-childRect.x(), -childRect.y());
+    }
+
+    legacyRootBox()->setX(0);
+    legacyRootBox()->setY(0);
+    legacyRootBox()->setLogicalWidth(childRect.width());
+    legacyRootBox()->setLogicalHeight(childRect.height());
+
+    auto boundingRect = enclosingLayoutRect(childRect);
+    legacyRootBox()->setLineTopBottomPositions(0, boundingRect.height(), 0, boundingRect.height());
+}
+
+static inline void swapItemsInLayoutAttributes(SVGTextLayoutAttributes* firstAttributes, SVGTextLayoutAttributes* lastAttributes, unsigned firstPosition, unsigned lastPosition)
+{
+    SVGCharacterDataMap::iterator itFirst = firstAttributes->characterDataMap().find(firstPosition + 1);
+    SVGCharacterDataMap::iterator itLast = lastAttributes->characterDataMap().find(lastPosition + 1);
+    bool firstPresent = itFirst != firstAttributes->characterDataMap().end();
+    bool lastPresent = itLast != lastAttributes->characterDataMap().end();
+    // We only want to perform the swap if both inline boxes are absolutely positioned.
+    if (!firstPresent || !lastPresent)
+        return;
+
+    std::swap(itFirst->value, itLast->value);
+}
+
+static inline void findFirstAndLastAttributesInVector(Vector<SVGTextLayoutAttributes*>& attributes, RenderSVGInlineText* firstContext, RenderSVGInlineText* lastContext, SVGTextLayoutAttributes*& first, SVGTextLayoutAttributes*& last)
+{
+    first = nullptr;
+    last = nullptr;
+
+    unsigned attributesSize = attributes.size();
+    for (unsigned i = 0; i < attributesSize; ++i) {
+        SVGTextLayoutAttributes* current = attributes[i];
+        if (!first && firstContext == &current->context())
+            first = current;
+        if (!last && lastContext == &current->context())
+            last = current;
+        if (first && last)
+            break;
+    }
+
+    ASSERT(first);
+    ASSERT(last);
+}
+
+WTF_ALLOW_UNSAFE_BUFFER_USAGE_BEGIN
+static inline void reverseInlineBoxRangeAndValueListsIfNeeded(Vector<SVGTextLayoutAttributes*>& attributes, Vector<InlineIterator::LeafBoxIterator>::iterator first, Vector<InlineIterator::LeafBoxIterator>::iterator last)
+{
+    // This is a copy of std::reverse(first, last). It additionally assures that the metrics map within the renderers belonging to the InlineBoxes are reordered as well.
+    while (true)  {
+        if (first == last || first == --last)
+            return;
+        auto* legacyFirst = (*first)->legacyInlineBox();
+        auto* legacyLast = (*last)->legacyInlineBox();
+        if (!is<SVGInlineTextBox>(legacyFirst) || !is<SVGInlineTextBox>(legacyLast)) {
+            auto temp = *first;
+            *first = *last;
+            *last = temp;
+            ++first;
+            continue;
+        }
+
+        auto& firstTextBox = downcast<SVGInlineTextBox>(*legacyFirst);
+        auto& lastTextBox = downcast<SVGInlineTextBox>(*legacyLast);
+
+        // Reordering is only necessary for BiDi text that is _absolutely_ positioned.
+        if (firstTextBox.len() == 1 && firstTextBox.len() == lastTextBox.len()) {
+            RenderSVGInlineText& firstContext = firstTextBox.renderer();
+            RenderSVGInlineText& lastContext = lastTextBox.renderer();
+
+            SVGTextLayoutAttributes* firstAttributes = nullptr;
+            SVGTextLayoutAttributes* lastAttributes = nullptr;
+            findFirstAndLastAttributesInVector(attributes, &firstContext, &lastContext, firstAttributes, lastAttributes);
+            swapItemsInLayoutAttributes(firstAttributes, lastAttributes, firstTextBox.start(), lastTextBox.start());
+        }
+
+        auto temp = *first;
+        *first = *last;
+        *last = temp;
+
+        ++first;
+    }
+}
+WTF_ALLOW_UNSAFE_BUFFER_USAGE_END
+
+void RenderSVGText::reorderValueListsToLogicalOrder()
+{
+    auto lineBox = InlineIterator::LineBoxIterator(legacyRootBox());
+    if (!lineBox)
+        return;
+
+    InlineIterator::leafBoxesInLogicalOrder(lineBox, [&](auto first, auto last) {
+        reverseInlineBoxRangeAndValueListsIfNeeded(m_layoutAttributes, first, last);
+    });
+
 }
 
 bool RenderSVGText::nodeAtFloatPoint(const HitTestRequest& request, HitTestResult& result, const FloatPoint& pointInParent, HitTestAction hitTestAction)
@@ -565,7 +772,7 @@ VisiblePosition RenderSVGText::positionForPoint(const LayoutPoint& pointInConten
         return createVisiblePosition(0, Affinity::Downstream);
 
     ASSERT(childrenInline());
-    LegacyInlineBox* closestBox = downcast<SVGRootInlineBox>(*rootBox).closestLeafChildForPosition(pointInContents);
+    LegacyInlineBox* closestBox = rootBox->closestLeafChildForPosition(pointInContents);
     if (!closestBox)
         return createVisiblePosition(0, Affinity::Downstream);
 
@@ -764,5 +971,11 @@ void RenderSVGText::styleDidChange(StyleDifference diff, const RenderStyle* oldS
 
     RenderSVGBlock::styleDidChange(diff, oldStyle);
 }
+
+SVGRootInlineBox* RenderSVGText::legacyRootBox() const
+{
+    return downcast<SVGRootInlineBox>(RenderSVGBlock::legacyRootBox());
+}
+
 
 }
