+++ Source/WebCore/rendering/svg/SVGRootInlineBox.cpp
@@ -25,16 +25,9 @@
 #include "config.h"
 #include "SVGRootInlineBox.h"
 
-#include "GraphicsContext.h"
-#include "InlineIteratorLogicalOrderTraversal.h"
 #include "RenderSVGText.h"
-#include "RenderSVGTextPath.h"
 #include "SVGElementTypeHelpers.h"
-#include "SVGInlineFlowBox.h"
 #include "SVGInlineTextBoxInlines.h"
-#include "SVGNames.h"
-#include "SVGRenderingContext.h"
-#include "SVGTextPositioningElement.h"
 #include <wtf/TZoneMallocInlines.h>
 
 namespace WebCore {
@@ -50,124 +43,6 @@ SVGRootInlineBox::SVGRootInlineBox(RenderSVGText& renderSVGText)
 RenderSVGText& SVGRootInlineBox::renderSVGText() const
 {
     return downcast<RenderSVGText>(blockFlow());
-}
-
-void SVGRootInlineBox::computePerCharacterLayoutInformation()
-{
-    auto& textRoot = downcast<RenderSVGText>(blockFlow());
-
-    Vector<SVGTextLayoutAttributes*>& layoutAttributes = textRoot.layoutAttributes();
-    if (layoutAttributes.isEmpty())
-        return;
-
-    if (textRoot.needsReordering())
-        reorderValueListsToLogicalOrder(layoutAttributes);
-
-    // Perform SVG text layout phase two (see SVGTextLayoutEngine for details).
-    SVGTextLayoutEngine characterLayout(layoutAttributes);
-    layoutCharactersInTextBoxes(this, characterLayout);
-
-    // Perform SVG text layout phase three (see SVGTextChunkBuilder for details).
-    auto fragmentMap = characterLayout.finishLayout();
-
-    // Perform SVG text layout phase four
-    // Position & resize all SVGInlineText/FlowBoxes in the inline box tree, resize the root box as well as the RenderSVGText parent block.
-    auto childRect = layoutChildBoxes(this, fragmentMap);
-    layoutRootBox(childRect);
-}
-
-void SVGRootInlineBox::layoutCharactersInTextBoxes(LegacyInlineFlowBox* start, SVGTextLayoutEngine& characterLayout)
-{
-    for (auto* child = start->firstChild(); child; child = child->nextOnLine()) {
-        if (auto* legacyTextBox = dynamicDowncast<SVGInlineTextBox>(*child)) {
-            ASSERT(is<RenderSVGInlineText>(legacyTextBox->renderer()));
-            characterLayout.layoutInlineTextBox(InlineIterator::svgTextBoxFor(legacyTextBox));
-        } else {
-            // Skip generated content.
-            RefPtr node = child->renderer().node();
-            if (!node)
-                continue;
-
-            auto& flowBox = downcast<SVGInlineFlowBox>(*child);
-            bool isTextPath = node->hasTagName(SVGNames::textPathTag);
-            if (isTextPath) {
-                // Build text chunks for all <textPath> children, using the line layout algorithm.
-                // This is needeed as text-anchor is just an additional startOffset for text paths.
-                SVGTextLayoutEngine lineLayout(characterLayout.layoutAttributes());
-                layoutCharactersInTextBoxes(&flowBox, lineLayout);
-
-                characterLayout.beginTextPathLayout(downcast<RenderSVGTextPath>(child->renderer()), lineLayout);
-            }
-
-            layoutCharactersInTextBoxes(&flowBox, characterLayout);
-
-            if (isTextPath)
-                characterLayout.endTextPathLayout();
-        }
-    }
-}
-
-FloatRect SVGRootInlineBox::layoutChildBoxes(LegacyInlineFlowBox* start, SVGTextFragmentMap& fragmentMap)
-{
-    FloatRect childRect;
-
-    for (auto* child = start->firstChild(); child; child = child->nextOnLine()) {
-        FloatRect boxRect;
-        if (auto* textBox = dynamicDowncast<SVGInlineTextBox>(*child)) {
-            ASSERT(is<RenderSVGInlineText>(textBox->renderer()));
-
-            auto it = fragmentMap.find(makeKey(*InlineIterator::svgTextBoxFor(textBox)));
-            if (it != fragmentMap.end())
-                textBox->setTextFragments(WTFMove(it->value));
-
-            boxRect = textBox->calculateBoundaries();
-            textBox->setX(boxRect.x());
-            textBox->setY(boxRect.y());
-            textBox->setLogicalWidth(boxRect.width());
-            textBox->setLogicalHeight(boxRect.height());
-        } else {
-            // Skip generated content.
-            if (!child->renderer().node())
-                continue;
-
-            auto& flowBox = downcast<SVGInlineFlowBox>(*child);
-            layoutChildBoxes(&flowBox, fragmentMap);
-
-            boxRect = flowBox.calculateBoundaries();
-            flowBox.setX(boxRect.x());
-            flowBox.setY(boxRect.y());
-            flowBox.setLogicalWidth(boxRect.width());
-            flowBox.setLogicalHeight(boxRect.height());
-        }
-        childRect.unite(boxRect);
-    }
-
-    return childRect;
-}
-
-void SVGRootInlineBox::layoutRootBox(const FloatRect& childRect)
-{
-    RenderSVGText& parentBlock = renderSVGText();
-
-    // Finally, assign the root block position, now that all content is laid out.
-    parentBlock.updatePositionAndOverflow(childRect);
-
-    // Position all children relative to the parent block.
-    for (auto* child = firstChild(); child; child = child->nextOnLine()) {
-        // Skip generated content.
-        if (!child->renderer().node())
-            continue;
-        child->adjustPosition(-childRect.x(), -childRect.y());
-    }
-
-    // Position ourselves.
-    setX(0);
-    setY(0);
-    setLogicalWidth(childRect.width());
-    setLogicalHeight(childRect.height());
-
-    auto boundingRect = enclosingLayoutRect(childRect);
-    setLineTopBottomPositions(0, boundingRect.height(), 0, boundingRect.height());
 }
 
 LegacyInlineBox* SVGRootInlineBox::closestLeafChildForPosition(const LayoutPoint& point)
@@ -192,90 +67,6 @@ LegacyInlineBox* SVGRootInlineBox::closestLeafChildForPosition(const LayoutPoint
     }
 
     return closestLeaf ? closestLeaf : lastLeaf;
-}
-
-static inline void swapItemsInLayoutAttributes(SVGTextLayoutAttributes* firstAttributes, SVGTextLayoutAttributes* lastAttributes, unsigned firstPosition, unsigned lastPosition)
-{
-    SVGCharacterDataMap::iterator itFirst = firstAttributes->characterDataMap().find(firstPosition + 1);
-    SVGCharacterDataMap::iterator itLast = lastAttributes->characterDataMap().find(lastPosition + 1);
-    bool firstPresent = itFirst != firstAttributes->characterDataMap().end();
-    bool lastPresent = itLast != lastAttributes->characterDataMap().end();
-    // We only want to perform the swap if both inline boxes are absolutely positioned.
-    if (!firstPresent || !lastPresent)
-        return;
-
-    std::swap(itFirst->value, itLast->value);
-}
-
-static inline void findFirstAndLastAttributesInVector(Vector<SVGTextLayoutAttributes*>& attributes, RenderSVGInlineText* firstContext, RenderSVGInlineText* lastContext,
-                                                      SVGTextLayoutAttributes*& first, SVGTextLayoutAttributes*& last)
-{
-    first = nullptr;
-    last = nullptr;
-
-    unsigned attributesSize = attributes.size();
-    for (unsigned i = 0; i < attributesSize; ++i) {
-        SVGTextLayoutAttributes* current = attributes[i];
-        if (!first && firstContext == &current->context())
-            first = current;
-        if (!last && lastContext == &current->context())
-            last = current;
-        if (first && last)
-            break;
-    }
-
-    ASSERT(first);
-    ASSERT(last);
-}
-
-WTF_ALLOW_UNSAFE_BUFFER_USAGE_BEGIN
-static inline void reverseInlineBoxRangeAndValueListsIfNeeded(Vector<SVGTextLayoutAttributes*>& attributes, Vector<InlineIterator::LeafBoxIterator>::iterator first, Vector<InlineIterator::LeafBoxIterator>::iterator last)
-{
-    // This is a copy of std::reverse(first, last). It additionally assures that the metrics map within the renderers belonging to the InlineBoxes are reordered as well.
-    while (true)  {
-        if (first == last || first == --last)
-            return;
-        auto* legacyFirst = (*first)->legacyInlineBox();
-        auto* legacyLast = (*last)->legacyInlineBox();
-        if (!is<SVGInlineTextBox>(legacyFirst) || !is<SVGInlineTextBox>(legacyLast)) {
-            auto temp = *first;
-            *first = *last;
-            *last = temp;
-            ++first;
-            continue;
-        }
-
-        auto& firstTextBox = downcast<SVGInlineTextBox>(*legacyFirst);
-        auto& lastTextBox = downcast<SVGInlineTextBox>(*legacyLast);
-
-        // Reordering is only necessary for BiDi text that is _absolutely_ positioned.
-        if (firstTextBox.len() == 1 && firstTextBox.len() == lastTextBox.len()) {
-            RenderSVGInlineText& firstContext = firstTextBox.renderer();
-            RenderSVGInlineText& lastContext = lastTextBox.renderer();
-
-            SVGTextLayoutAttributes* firstAttributes = nullptr;
-            SVGTextLayoutAttributes* lastAttributes = nullptr;
-            findFirstAndLastAttributesInVector(attributes, &firstContext, &lastContext, firstAttributes, lastAttributes);
-            swapItemsInLayoutAttributes(firstAttributes, lastAttributes, firstTextBox.start(), lastTextBox.start());
-        }
-
-        auto temp = *first;
-        *first = *last;
-        *last = temp;
-
-        ++first;
-    }
-}
-WTF_ALLOW_UNSAFE_BUFFER_USAGE_END
-
-void SVGRootInlineBox::reorderValueListsToLogicalOrder(Vector<SVGTextLayoutAttributes*>& attributes)
-{
-    auto lineBox = InlineIterator::LineBoxIterator(this);
-
-    InlineIterator::leafBoxesInLogicalOrder(lineBox, [&](auto first, auto last) {
-        reverseInlineBoxRangeAndValueListsIfNeeded(attributes, first, last);
-    });
-
 }
 
 } // namespace WebCore
